+++ price-feeder/oracle/provider/gate.go
@@ -13,7 +13,7 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/rs/zerolog"
 
-	"github.com/umee-network/umee/price-feeder/oracle/types"
+	"github.com/umee-network/umee/price-feeder/v2/oracle/types"
 )
 
 const (
@@ -248,7 +248,7 @@ func (p *GateProvider) getTickerPrice(cp types.CurrencyPair) (types.TickerPrice,
 	return types.TickerPrice{}, fmt.Errorf("gate failed to get ticker price for %s", gp)
 }
 
-func (p *GateProvider) messageReceived(messageType int, bz []byte) {
+func (p *GateProvider) messageReceived(_ int, bz []byte) {
 	var (
 		gateEvent GateEvent
 		gateErr   error
@@ -262,7 +262,6 @@ func (p *GateProvider) messageReceived(messageType int, bz []byte) {
 		case "success":
 			return
 		case "":
-			break
 		default:
 			return
 		}