+++ price-feeder/oracle/provider/okx.go
@@ -13,7 +13,7 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/rs/zerolog"
 
-	"github.com/umee-network/umee/price-feeder/oracle/types"
+	"github.com/umee-network/umee/price-feeder/v2/oracle/types"
 )
 
 const (
@@ -244,7 +244,7 @@ func (p *OkxProvider) getCandlePrices(cp types.CurrencyPair) ([]types.CandlePric
 	return candleList, nil
 }
 
-func (p *OkxProvider) messageReceived(messageType int, bz []byte) {
+func (p *OkxProvider) messageReceived(_ int, bz []byte) {
 	var (
 		tickerResp OkxTickerResponse
 		tickerErr  error