+++ price-feeder/oracle/provider/osmosisv2.go
@@ -12,7 +12,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/gorilla/websocket"
 	"github.com/rs/zerolog"
-	"github.com/umee-network/umee/price-feeder/oracle/types"
+	"github.com/umee-network/umee/price-feeder/v2/oracle/types"
 )
 
 const (
@@ -189,7 +189,7 @@ func (p *OsmosisV2Provider) getCandlePrices(key string) ([]types.CandlePrice, er
 	return candleList, nil
 }
 
-func (p *OsmosisV2Provider) messageReceived(messageType int, bz []byte) {
+func (p *OsmosisV2Provider) messageReceived(_ int, bz []byte) {
 	// check if message is an ack first
 	if string(bz) == "ack" {
 		return