+++ x/oracle/keeper/genesis_test.go
@@ -9,7 +9,7 @@ import (
 func (s *IntegrationTestSuite) TestIterateAllHistoricPrices() {
 	keeper, ctx := s.app.OracleKeeper, s.ctx
 
-	historicPrices := []types.HistoricPrice{
+	historicPrices := []types.Price{
 		{BlockNum: 10, ExchangeRateTuple: types.ExchangeRateTuple{
 			Denom: "umee", ExchangeRate: sdk.MustNewDecFromStr("20.45"),
 		}},
@@ -28,10 +28,10 @@ func (s *IntegrationTestSuite) TestIterateAllHistoricPrices() {
 		keeper.SetHistoricPrice(ctx, hp.ExchangeRateTuple.Denom, hp.BlockNum, hp.ExchangeRateTuple.ExchangeRate)
 	}
 
-	newPrices := []types.HistoricPrice{}
+	newPrices := []types.Price{}
 	keeper.IterateAllHistoricPrices(
 		ctx,
-		func(historicPrice types.HistoricPrice) bool {
+		func(historicPrice types.Price) bool {
 			newPrices = append(newPrices, historicPrice)
 			return false
 		},
@@ -62,13 +62,13 @@ func (s *IntegrationTestSuite) TestIterateAllMedianPrices() {
 	}
 
 	for _, m := range medians {
-		keeper.SetMedian(ctx, m.Denom, m.ExchangeRate)
+		keeper.SetHistoricMedian(ctx, m.Denom, uint64(s.ctx.BlockHeight()), m.ExchangeRate)
 	}
 
-	newMedians := []types.ExchangeRateTuple{}
+	newMedians := []types.Price{}
 	keeper.IterateAllMedianPrices(
 		ctx,
-		func(median types.ExchangeRateTuple) bool {
+		func(median types.Price) bool {
 			newMedians = append(newMedians, median)
 			return false
 		},
@@ -78,8 +78,8 @@ func (s *IntegrationTestSuite) TestIterateAllMedianPrices() {
 FOUND:
 	for _, oldMedian := range medians {
 		for _, newMedian := range newMedians {
-			if oldMedian.Denom == newMedian.Denom {
-				s.Require().Equal(oldMedian.ExchangeRate, newMedian.ExchangeRate)
+			if oldMedian.Denom == newMedian.ExchangeRateTuple.Denom {
+				s.Require().Equal(oldMedian.ExchangeRate, newMedian.ExchangeRateTuple.ExchangeRate)
 				continue FOUND
 			}
 		}
@@ -96,13 +96,13 @@ func (s *IntegrationTestSuite) TestIterateAllMedianDeviationPrices() {
 	}
 
 	for _, m := range medians {
-		keeper.SetMedianDeviation(ctx, m.Denom, m.ExchangeRate)
+		keeper.SetHistoricMedianDeviation(ctx, m.Denom, uint64(s.ctx.BlockHeight()), m.ExchangeRate)
 	}
 
-	newMedians := []types.ExchangeRateTuple{}
+	newMedians := []types.Price{}
 	keeper.IterateAllMedianDeviationPrices(
 		ctx,
-		func(median types.ExchangeRateTuple) bool {
+		func(median types.Price) bool {
 			newMedians = append(newMedians, median)
 			return false
 		},
@@ -112,8 +112,8 @@ func (s *IntegrationTestSuite) TestIterateAllMedianDeviationPrices() {
 FOUND:
 	for _, oldMedian := range medians {
 		for _, newMedian := range newMedians {
-			if oldMedian.Denom == newMedian.Denom {
-				s.Require().Equal(oldMedian.ExchangeRate, newMedian.ExchangeRate)
+			if oldMedian.Denom == newMedian.ExchangeRateTuple.Denom {
+				s.Require().Equal(oldMedian.ExchangeRate, newMedian.ExchangeRateTuple.ExchangeRate)
 				continue FOUND
 			}
 		}