+++ x/oracle/keeper/migrations.go
@@ -2,8 +2,6 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/umee-network/umee/v3/x/oracle/types"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -18,9 +16,9 @@ func NewMigrator(keeper *Keeper) Migrator {
 
 // Migrate1to2 migrates from version 1 to 2.
 func (m Migrator) Migrate1to2(ctx sdk.Context) error {
-	m.keeper.SetStampPeriod(ctx, 1)
-	m.keeper.SetPrunePeriod(ctx, 1)
-	m.keeper.SetMedianPeriod(ctx, 1)
-	m.keeper.SetHistoricAcceptList(ctx, types.DenomList{})
+	m.keeper.SetHistoricStampPeriod(ctx, 1)
+	m.keeper.SetMedianStampPeriod(ctx, 1)
+	m.keeper.SetMaximumPriceStamps(ctx, 1)
+	m.keeper.SetMaximumMedianStamps(ctx, 1)
 	return nil
 }